requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.8
importlib_metadata==8.4.0
Markdown==3.7
outcome==1.3.0.post0
psycopg==3.2.1
PySocks==1.7.1
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.20.1
